arp now working

+++ parsers.h
@@ -8,7 +8,7 @@ struct  __attribute__((packed)) arp_pac {
     uint16_t ptype;
     uint8_t hlen;
     uint8_t plen;
-    uint8_t oper;
+    uint16_t oper;
     uint8_t sha[6];
     uint8_t spa[4];
     uint8_t tha[6];
